requirements.txt
matplotlib==3.10.0
seaborn==0.13.2
tqdm==4.67.1
scikit-learn==1.6.1
pandas==2.2.2
numpy==2.0.2
tensorflow==2.19.0
keras==3.13.1
pillow==11.3.0